// src/app/model/wallet/merchant-setting/merchant-setting.ts
export class MerchantSetting {
    url_cancel: string;
    url_return: string;
    url_notify: string;
    url_logo: string;
    merchantName: string;
    hashCheckSum: boolean;
    uniqueOrderId: boolean;
    vcnAgent: boolean;
    requireOTPVcn: boolean;
    autocapture: boolean;
    silentReturn: boolean;
    sendMail: boolean;
}